requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-phonenumber-field==5.0.0
djangorestframework==3.12.1
flake8==3.8.4
gunicorn==20.0.4
install==1.3.4
mccabe==0.6.1
phonenumbers==8.12.11
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-decouple==3.3
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
